requirements.txt
django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-decouple==3.8
pytest==8.1.1
pytest-django==4.8.0
gunicorn==22.0.0
django-cors-headers==4.3.1
setuptools==70.0.0